# packages/BarryKit/Sources/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Barry API",
    "version": "1.0.0",
    "description": "The single supported HTTP contract for Barry applications."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/sessions/{sessionId}/diff": {
      "get": {
        "operationId": "getSessionDiff",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-log": {
      "get": {
        "operationId": "getSessionGitLog",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitLogResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-status": {
      "get": {
        "operationId": "getSessionGitStatus",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatusResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-commit": {
      "post": {
        "operationId": "commitSessionChanges",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommitResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-push": {
      "post": {
        "operationId": "pushSessionChanges",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitPushResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-branches": {
      "get": {
        "operationId": "listSessionBranches",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoBranchesResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-switch-branch": {
      "post": {
        "operationId": "switchSessionBranch",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitBranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/git-create-branch": {
      "post": {
        "operationId": "createSessionBranch",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitBranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/changes": {
      "get": {
        "operationId": "getSessionChangeSummary",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionChangeSummaryResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      },
...
      "DiffResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "repoPath": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "baseBranch": {
            "type": "string"
          },
          "currentBranch": {
            "type": "string"
          },
          "commit": {
            "type": "string"
          },
          "onMainBranch": {
            "type": "boolean"
          },
          "refOnly": {
            "type": "boolean"
          },
          "diff": {
            "type": "string"
          },
          "hasStagedChanges": {
            "type": "boolean"
          },
          "hasUnstagedChanges": {
            "type": "boolean"
          },
          "hasUntrackedFiles": {
            "type": "boolean"
          }
        },
        "required": [
          "repoPath",
          "mode",
          "diff"
        ],
        "additionalProperties": false
      },
      "GitLogResponse": {
        "type": "object",
        "properties": {
          "repoPath": {
            "type": "string"
          },
          "baseBranch": {
            "type": "string"
          },
          "currentBranch": {
            "type": "string"
          },
          "commits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "hash": {
                  "type": "string"
                },
                "shortHash": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "author": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "filesChanged": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "insertions": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "deletions": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "hash",
                "shortHash",
                "subject",
                "author",
                "date",
                "filesChanged",
                "insertions",
                "deletions"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "repoPath",
          "baseBranch",
          "currentBranch",
          "commits"
        ],
        "additionalProperties": false
      },
...
      "GitStatusResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "repoPath": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "staged": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "unstaged": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "untracked": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hasUpstream": {
            "type": "boolean"
          },
          "ahead": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "behind": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "sessionId",
          "repoPath",
          "branch",
          "staged",
          "unstaged",
          "untracked",
          "conflicts",
          "hasUpstream",
          "ahead",
          "behind"
        ],
        "additionalProperties": false
      },
      "GitCommitRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": false
      },
      "GitCommitResponse": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          }
        },
        "required": [
          "hash"
        ],
        "additionalProperties": false
      },
      "GitPushResponse": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string"
          }
        },
        "required": [
          "branch"
        ],
        "additionalProperties": false
      },
      "GitBranchRequest": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "branch"
        ],
        "additionalProperties": false
      },
...
    }
  }
}